characteristic/u2af8/src/main/java/org/im97mori/ble/characteristic/u2af8/FixedString8.java
package org.im97mori.ble.characteristic.u2af8;


import org.im97mori.ble.ByteArrayInterface;

import androidx.annotation.NonNull;

/**
 * Fixed String 8 (Characteristics UUID: 0x2AF8)
 */
public class FixedString8 implements ByteArrayInterface {

    /**
     * Fixed String
     */
    private final String mFixedString;

    /**
     * Constructor from byte array
     *
     * @param values byte array from <a href="https://developer.android.com/reference/android/bluetooth/BluetoothGattCharacteristic#getValue()">BluetoothGattCharacteristic#getValue()</a>
     */
    public FixedString8(@NonNull byte[] values) {
        mFixedString = new String(values, 0, 8);
    }

    /**
     * Constructor from parameters
     * 
     * @param fixedString Fixed String
     */
    public FixedString8(@NonNull String fixedString) {
        mFixedString = new String(fixedString.getBytes(), 0, 8);
    }

    /**
     * @return Fixed String
     */
    public String getFixedString() {
        return mFixedString;
    }

    /**
     * {@inheritDoc}
     */
    @Override
    @NonNull
    public byte[] getBytes() {
        return mFixedString.getBytes();
    }

}
